certificate_system/requirements.txt
streamlit>=1.35.0
bcrypt>=4.1.2
sqlmodel>=0.0.14
sqlite3>=2.6.0
pandas>=2.2.2
openpyxl>=3.1.5
python-multipart>=0.0.9
pillow>=10.3.0
pdf2image>=1.17.0
requests>=2.32.3
# 新增文件处理依赖
python-magic-bin>=0.4.27  # Windows文件类型校验
pathlib>=1.0.1            # 路径处理